mutualize viscode_to_session
